Fix regression - import transaction contain failure inflation ops (#11)
* Patch Network fee charged should be 100 Stroops + 0.45%
* Patch LedgerHeader with basepercentagefee & maxFee
* Patch TransactionV0 fee to uint64
* Patch feepool account

+++ src/transactions/TransactionBridge.cpp
@@ -108,10 +108,10 @@ setSeqNum(TransactionFramePtr tx, int64_t seq)
 }
 
 void
-setFee(TransactionFramePtr tx, uint32_t fee)
+setFee(TransactionFramePtr tx, uint64_t fee)
 {
     auto& env = tx->getEnvelope();
-    uint32_t& f =
+    uint64_t& f =
         env.type() == ENVELOPE_TYPE_TX_V0 ? env.v0().tx.fee : env.v1().tx.fee;
     f = fee;
 }
